requirements.txt
Flask==2.3.3
openai==1.3.5
twilio==8.10.0
python-dotenv==1.0.0
google-api-python-client==2.110.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pytz==2023.3
APScheduler==3.10.4